fix: Change static directory name feat: Change UI for the mobile page and icon for the mouse
Graffiti themed with minimal js has been used for mobile_page
